app/requirements.txt
mysqlclient==2.1.1
Django==4.0.5
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-filter==22.1
django-url-filter==0.3.15
debugpy==1.6.0
msal==1.22.0
tzdata==2022.1
requests==2.28.0
pysftp==0.2.9
django-ses==3.1.0
django-storages==1.12.3
boto3==1.24.33
python-dotenv==0.20.0
python-dateutil==2.8.2
PyYAML==6.0
drf_yasg==1.21.7

